refactor(questions): extract V2 read helper to its own module
The V2 question-read logic (schema, parser, warn-once state, fallback
classification) was inlined in client.ts as a private method. Move it to
its own module (questions-v2.ts) so the directory-bootstrap caller
(#3288 / unit-5) can share one authoritative implementation instead of
duplicating the logic on its own branch.

The OpencodeService read path keeps its fetchPendingQuestionsV2 facade so
listPendingQuestions' V2-then-V1 sequence still reads naturally; the
facade now delegates to listPendingQuestionsViaV2 in the new module.

Also covers the 200-with-non-array-payload branch the helper already
handles (Array.isArray check) but was previously untested, and adds a
closed NonArrayPayload union type so the new test mocks the contract
drift explicitly.
